fix(#3060): split grip_modelling_tab.py — extract model data and panel widgets
- PressureVisualizationWidget and ContactMetricsWidget extracted to grip_plot_panel.py (207 lines)
- Scene XML construction helpers extracted to grip_xml_builder.py via GripSceneXmlBuilder (303 lines)
- grip_modelling_tab.py thinned from 1034 to 574 lines as orchestrator; backward-compat re-exports preserved
